Improve animation for better instant smoothness Enhance stress testing for improved results - Addressed a bug related to ball speed and emitted events Rearrange code for better organization Changes to be committed: 	modified:   backend/src/modules/ping-pong/Room.ts 	modified:   backend/src/modules/ping-pong/achievement.service.ts 	renamed:    backend/src/modules/ping-pong/types/Common.ts -> backend/src/modules/ping-pong/common/Common.ts 	renamed:    backend/src/modules/ping-pong/logic/Ball.ts -> backend/src/modules/ping-pong/component/Ball.ts 	renamed:    backend/src/modules/ping-pong/logic/Game.ts -> backend/src/modules/ping-pong/component/Game.ts 	renamed:    backend/src/modules/ping-pong/logic/Player.ts -> backend/src/modules/ping-pong/component/Player.ts 	modified:   backend/src/modules/ping-pong/ping-pong.gateway.ts 	modified:   backend/src/modules/ping-pong/ping-pong.module.ts 	modified:   backend/src/modules/ping-pong/ping-pong.service.ts 	modified:   frontend/app/(game)/ping-pong/(proceed-game)/service/Service.tsx 	modified:   frontend/app/(game)/ping-pong/context/WebSocketContext.tsx 	modified:   frontend/app/(game)/ping-pong/game/Ball.tsx 	modified:   frontend/app/(game)/ping-pong/game/Player.tsx

+++ backend/src/modules/ping-pong/ping-pong.gateway.ts
@@ -7,14 +7,17 @@ import {
 	WebSocketServer,
 	ConnectedSocket,
 } from '@nestjs/websockets';
+
 import { Server, Socket } from 'socket.io';
 
-import Room from "./Room";
-import { Mode, PlayerType } from './types/Common';
-import { PrismaService } from '../prisma/prisma.service';
+import Room from "src/modules/ping-pong/Room";
+import { PingPongService } from 'src/modules/ping-pong/ping-pong.service';
+import { Mode, PlayerType } from 'src/modules/ping-pong/common/Common';
+
+import { PrismaService } from 'src/modules/prisma/prisma.service';
+
 import { GlobalHelperService } from 'src/common/services/global_helper.service';
 import { SocketService } from 'src/common/services/socket.service';
-import { PingPongService } from './ping-pong.service';
 
 // namespace for websocket events (client -> server)
 @WebSocketGateway(
@@ -44,6 +47,7 @@ export default class PingPongGateway implements OnGatewayInit, OnGatewayConnecti
 
 		if (userId === undefined) {
 			this.server.to(client.id).emit('error', { error: 'Invalid Access Token' });
+			//! route to authentication page
 			client.disconnect();
 			return;
 		}
@@ -56,6 +60,12 @@ export default class PingPongGateway implements OnGatewayInit, OnGatewayConnecti
 
 	async handleDisconnect(client: Socket) {
 		const userId = this.socketService.getUserId(client.id, 'ping-pong');
+
+		if (userId === undefined) {
+			this.server.to(client.id).emit('error', { error: 'Invalid Access Token' });
+			//! route to authentication page
+			return;
+		}
 
 		console.log('DELETE CONNECTION: ' + client.id + ' ' + userId);
 		if (this.rooms.deletePlayerRoom(userId.toString())) {
@@ -138,6 +148,7 @@ export default class PingPongGateway implements OnGatewayInit, OnGatewayConnecti
 
 		if (userId === undefined) {
 			this.server.to(client.id).emit('error', { error: 'Invalid Access Token' });
+			//! route to authentication page
 			return undefined;
 		}
